requirements.txt
asgiref==3.5.2
Django==3.2.5
django-js-asset==2.0.0
django-tinymce==2.8.0
PyMySQL==1.0.2
pytz==2022.6
sqlparse==0.4.3
